requirements.txt
# Dependencias Python

python-docx>=0.8.11
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
python-dotenv>=1.0.0
